+++ responsibly/fairness/metrics/utils.py
@@ -1,3 +1,4 @@
+import pandas as pd
 from sklearn.metrics.classification import _check_targets
 
 
@@ -10,3 +11,10 @@ def _assert_binary(y1, y2=None):
 
     if y_type != 'binary':
         raise ValueError('y_true and y_pred must be binary.')
+
+
+def _groupby_y_x_sens(y_true, y_score, x_sens):
+    return (pd.DataFrame({'y_true': y_true,
+                          'y_score': y_score,
+                          'x_sens': x_sens})
+            .groupby('x_sens'))
